Fixed a few warnings. Updated project settings

+++ IsaacsIOSLibrary/Categories/NSDate/NSDate+Compare.h
@@ -14,6 +14,6 @@
 - (bool)isAfterDate:(NSDate*)otherDate;
 - (bool)isToday;
 - (bool)isYesterday;
-- (int)daysSince:(NSDate*)other;
+- (NSInteger)daysSince:(NSDate*)other;
 
 @end
